Move platform-specific code behind abstract interface
