+++ makefile
@@ -214,7 +214,7 @@ ALL_SOURCE_FILES = \
  lib/material.cpp lib/openglvis.cpp lib/palettes.cpp lib/sdl.cpp \
  lib/threads.cpp lib/vsdata.cpp lib/vssolution.cpp lib/vssolution3d.cpp \
  lib/vsvector.cpp lib/vsvector3d.cpp
-OBJC_SOURCE_FILES = $(if $(NOTMAC),,lib/sdl_mac.m)
+OBJC_SOURCE_FILES = $(if $(NOTMAC),,lib/sdl_mac.mm)
 DESKTOP_ONLY_SOURCE_FILES = lib/gl/renderer_ff.cpp lib/threads.cpp lib/gl2ps.c
 WEB_ONLY_SOURCE_FILES = lib/aux_js.cpp
 LOGO_FILE = share/logo.rgba
@@ -227,14 +227,15 @@ HEADER_FILES = \
  lib/gl/attr_traits.hpp lib/gl/platform_gl.hpp lib/gl/renderer.hpp \
  lib/gl/renderer_core.hpp lib/gl/renderer_ff.hpp lib/gl/types.hpp \
  lib/aux_vis.hpp lib/font.hpp lib/gl2ps.h lib/logo.hpp lib/material.hpp \
- lib/openglvis.hpp lib/palettes.hpp lib/sdl.hpp lib/sdl_mac.hpp \
- lib/threads.hpp lib/visual.hpp lib/vsdata.hpp lib/vssolution.hpp \
- lib/vssolution3d.hpp lib/vsvector.hpp lib/vsvector3d.hpp
+ lib/openglvis.hpp lib/palettes.hpp lib/sdl.hpp lib/sdl_helper.hpp \
+ lib/sdl_mac.hpp lib/sdl_x11.hpp lib/threads.hpp lib/visual.hpp \
+ lib/vsdata.hpp lib/vssolution.hpp lib/vssolution3d.hpp lib/vsvector.hpp \
+ lib/vsvector3d.hpp
 
 DESKTOP_SOURCE_FILES = $(COMMON_SOURCE_FILES) $(DESKTOP_ONLY_SOURCE_FILES) $(LOGO_FILE_CPP)
 WEB_SOURCE_FILES     = $(COMMON_SOURCE_FILES) $(WEB_ONLY_SOURCE_FILES)
 OBJECT_FILES1        = $(DESKTOP_SOURCE_FILES:.cpp=.o)
-OBJECT_FILES         = $(OBJECT_FILES1:.c=.o) $(OBJC_SOURCE_FILES:.m=.o)
+OBJECT_FILES         = $(OBJECT_FILES1:.c=.o) $(OBJC_SOURCE_FILES:.mm=.o)
 BYTECODE_FILES       = $(WEB_SOURCE_FILES:.cpp=.bc)
 
 # Targets
@@ -246,8 +247,8 @@ BYTECODE_FILES       = $(WEB_SOURCE_FILES:.cpp=.bc)
 %.o: %.c %.h
 	$(Ccc) -o $@ -c $<
 
-%.o: %.m
-	$(CC) $(CFLAGS) -o $@ -c $<
+%.o: %.mm
+	$(CCC) -o $@ -c $<
 
 %.bc: %.cpp
 	$(EMCC) $(EMCC_OPTS) -c $< -o $@
